LayoutTests/inspector/codemirror/prettyprinting-css-rules.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/prettyprinting/utilities.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("CodeMirror.PrettyPrinting.CSSRules");

    addPrettyPrintingTests(suite, "css-rule", [
        "resources/prettyprinting/css-rule-tests/invalid-property-is-not-removed.css",
        "resources/prettyprinting/css-rule-tests/remove-newline-between-values.css",
        "resources/prettyprinting/css-rule-tests/remove-whitespace-before-colon.css",
        "resources/prettyprinting/css-rule-tests/remove-whitespace-before-semicolon.css",
        "resources/prettyprinting/css-rule-tests/remove-whitespace-before-property.css",
        "resources/prettyprinting/css-rule-tests/remove-whitespace-before-prefixed-property.css",
        "resources/prettyprinting/css-rule-tests/remove-whitespace-before-invalid-property.css",
        "resources/prettyprinting/css-rule-tests/remove-whitespace-before-comment.css",
        "resources/prettyprinting/css-rule-tests/split-comment-followed-by-property.css",
        "resources/prettyprinting/css-rule-tests/split-comment-followed-by-prefixed-property.css",
        "resources/prettyprinting/css-rule-tests/split-comment-followed-by-invalid-property.css",
        "resources/prettyprinting/css-rule-tests/split-comment-followed-by-comment.css",
        "resources/prettyprinting/css-rule-tests/split-property-followed-by-property.css",
        "resources/prettyprinting/css-rule-tests/split-property-followed-by-prefixed-property.css",
        "resources/prettyprinting/css-rule-tests/split-property-followed-by-invalid-property.css",
        "resources/prettyprinting/css-rule-tests/split-property-followed-by-comment.css",
        "resources/prettyprinting/css-rule-tests/split-invalid-property-followed-by-property.css",
        "resources/prettyprinting/css-rule-tests/split-invalid-property-followed-by-prefixed-property.css",
        "resources/prettyprinting/css-rule-tests/split-invalid-property-followed-by-invalid-property.css",
        "resources/prettyprinting/css-rule-tests/split-invalid-property-followed-by-comment.css",
        "resources/prettyprinting/css-rule-tests/split-property-without-semicolon-followed-by-comment-and-property.css",
        "resources/prettyprinting/css-rule-tests/add-whitespace-between-values.css",
        "resources/prettyprinting/css-rule-tests/add-whitespace-between-rules.css",
        "resources/prettyprinting/css-rule-tests/add-whitespace-after-colon.css",
        "resources/prettyprinting/css-rule-tests/add-whitespace-after-comma.css",
        "resources/prettyprinting/css-rule-tests/do-not-append-semicolon.css",
        "resources/prettyprinting/css-rule-tests/do-not-add-whitespace-before-prefixed-property-value.css",
        "resources/prettyprinting/css-rule-tests/keep-prefixed-value.css",
        "resources/prettyprinting/css-rule-tests/css-variables.css",
    ]);

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test CSS Rule pretty printing with CodeMirror.</p>
</body>
</html>
