LayoutTests/inspector/codemirror/prettyprinting-css-rules-expected.txt

Test CSS Rule pretty printing with CodeMirror.


== Running test suite: CodeMirror.PrettyPrinting.CSSRules
-- Running test case: CodeMirror.PrettyPrinting.CSSRules.invalid-property-is-not-removed.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.remove-newline-between-values.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.remove-whitespace-before-colon.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.remove-whitespace-before-semicolon.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.remove-whitespace-before-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.remove-whitespace-before-prefixed-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.remove-whitespace-before-invalid-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.remove-whitespace-before-comment.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-comment-followed-by-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-comment-followed-by-prefixed-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-comment-followed-by-invalid-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-comment-followed-by-comment.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-property-followed-by-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-property-followed-by-prefixed-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-property-followed-by-invalid-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-property-followed-by-comment.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-invalid-property-followed-by-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-invalid-property-followed-by-prefixed-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-invalid-property-followed-by-invalid-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-invalid-property-followed-by-comment.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.split-property-without-semicolon-followed-by-comment-and-property.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.add-whitespace-between-values.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.add-whitespace-between-rules.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.add-whitespace-after-colon.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.add-whitespace-after-comma.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.do-not-append-semicolon.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.do-not-add-whitespace-before-prefixed-property-value.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.keep-prefixed-value.css
PASS

-- Running test case: CodeMirror.PrettyPrinting.CSSRules.css-variables.css
PASS
